Create README.md
Objective: Proposal for optimizing financial resources in "Plano ABC+" (Sectoral Plan for Adaptation to Climate Change and Low Carbon Emissions in Brazilian Agriculture, with a view to Sustainable Development: 2020-2030).

Climate change poses urgent challenges to global agriculture, with the agricultural sector being one of the main emitters of greenhouse gases (GHG). In Brazil, the ABC and ABC+ Plans represent strategic initiatives to promote sustainable agricultural practices and reduce GHG emissions. This article aims to analyze the optimization of financial resource allocation within these plans, seeking to maximize the efficiency of sustainable practices and contribute to the development of a more resilient agricultural system. Multi-objective optimization methods and rule-based heuristics will be applied to allocate resources, taking into account the specificities of the Brazilian context. Furthermore, opportunities and challenges in implementing these practices in the agricultural sector will be discussed as tools to encourage the adoption of sustainable practices and the reduction of GHG emissions. The findings of this research will contribute to the enhancement of public policies promoting sustainable agriculture in Brazil and to achieving the climate change mitigation goals outlined in the Paris Agreement.